django_be/requirements.txt
appnope==0.1.3
asgiref==3.7.2
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.3
black==23.3.0
blinker==1.6.2
bs4==0.0.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.3
comm==0.1.3
cryptography==41.0.7
debugpy==1.6.7
decorator==5.1.1
distlib==0.3.7
Django==5.0.1
django-cors-headers==4.3.1
djangorestframework==3.14.0
executing==1.2.0
filelock==3.13.1
flake8==6.0.0
Flask==2.3.2
gnureadline==8.1.2
idna==3.6
ipykernel==6.22.0
ipython==8.13.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.2.0
jupyter_core==5.3.0
Mako==1.3.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mmh3==4.0.1
mypy-extensions==1.0.0
nest-asyncio==1.5.6
openai==1.11.1
packaging==23.1
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==4.0.0
prettytable==3.9.0
prompt-toolkit==3.0.38
protobuf==4.21.12
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
Pygments==2.15.1
pyOpenSSL==23.3.0
pyotp==2.9.0
PyPDF2==3.0.1
pypng==0.20220715.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.4
PyYAML==6.0.1
pyzmq==25.0.2
qrcode==7.4.2
requests==2.31.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.8.0
urllib3==2.1.0
virtualenv==20.25.0
wcwidth==0.2.6
Werkzeug==2.3.6
pymongo==4.6.1
retrying==1.3.4
docx2txt==0.8
nltk==3.8.1